Add 'smooth muscle' as a broad synonym of 'visceral muscle'.
The literature on vertebrates frequently use 'smooth muscle'
interchangeably with 'visceral muscle', so we add that as a
synonym -- with a comment to warn that such an usage is
potentially misleading in flies as their visceral muscles are
striated.

We don't add a corresponding 'striated muscle' synomym to
'somatic muscle'. Using 'striated muscle' for 'somatic muscle'
seems infrequent, likely because of the risk of confusion with
cardiac muscles (which are also striated, but not somatic).