Review definitions of adult EE cells.
Add manual definitions to the different subtypes of EE cells to
state which other hormone(s) they express in addition to the
class-defining hormone.

Update the definition of the adult class II enteroendocrine cell
to mention that they co-express varying hormones depending on the
compartment (as for class I EE cells).