Add classes for larval Dh31 expressing cells.
Add a 'Dh31 cell of the embryonic/larval midgut' class covering all
Dh31-expressing cells in the larval midgut.

Add another more specific class for the particular subset of those cells
found at the junction between the anterior midgut and the middle midgut,
and shown to be required for peristalsis (Lajeunesse et al., 2010).