Add specific TF data for some optic lobe neurons.
Complement the definition of some of the optic lobe neurons with a
specific transcription factor that uniquely marks that type of neurons
among all the optic lobe neurons, based on the data from Özel et al.
(2021).

(These authors suggest many more potential markers, but most of them
mark several distinct populations and must be used in combinations to
label a specific type; here, only the markers specific to one type have
been added, except for the Fer2 TF which marks both the Lawf1 and Lawf2
neurons.)