Fix running FlyBase imports on M1 machines.
FlyBase imports use a Python script that requires psycopg2, which is not
present in the ODK container. It is normally installed through pip prior
to running the script, but on M1-powered machines this does not work
because as of now there is no pre-compiled binary package targeting the
arm64 architecture for psycopg2 in the PYPI repository.

The workaround here is to install psycopg2 through the underlying Ubuntu
distribution (which does provide a arm64 binary package for psycopg2)
via apt-get rather than through pip.

This allows running the import script on M1-powered machines and it
should not change anything for folks using Intel-powered ones.