Update leg stretch receptor neuron.
Add subclasses for the different stretch receptor neurons (for the
femur-tibia and the tibia-tarsus joints and for the pro-, meso-, and
metathoracic legs).

Add a class for the newly described stretch receptor of the prothoracic
coxa-trochanter joint (Kuan et al., 2020) (FBrf0247324).

Add expression of stum (FBgn0050263) in the stretch receptor neurons of
the femur-tibia and tibia-tarsus joints (Desai et al., 2014)
(FBrf0224410).