Fix definition of 'apodeme'.
The definition of 'apodeme' referred to an 'exoskeleton', which can only
be a mistake. The term was also classed as 'part of some endocuticle',
(because it's an 'endocuticular specialization') but 'endocuticle' is
defined as only existing in the pupa while apodemes exist in the adult,
so re-classify it as a 'cuticular specialization'.

This necessitated to also re-classify 'phallapodeme' as 'part of some
phallic sclerite' instead of 'is a phallic sclerite', otherwise the
phallapodeme ended up being both a 'acellular anatomical structure' and
a 'multicellular structure'.