Update definition of adult enteroendocrine cell.
Update the definition of 'adult enteroendocrine cell' to give the number
of such cells (from Chen et al., 2016 -- FBrf0232200) and mention their
origin from the EEP cells (already stated as a relationship, but not in
the definition).

Also add a 'adult enteroendocrine progenitor cell' synonym to 'adult
enteroendocrine precursor cell. Different authors use one term or the
other but they refer to the same cell type.