Rename Lccn1 neurons to LPLC.
The lobula complex columnar neurons type 1 (Lccn1) described in
Drosophila by Fischbach and Dittrich (1989) are the same as the lobula
plate-lobula (columnar) neurons (LPL or LPLC) later described in other
diptera by Strausfeld and Gilbert (1992). Since the term LPLC is favored
in recent literature, we rename that class to reflect usage. The
definition is also updated based on more recent descriptions of these
neurons.