Rename Lccn2 neuron to LLPC and broaden definition.
The lobula complex columnar neurons type 2 (Lccn2) described by
Fischbach and Dittrich (1989) are the same as those later described as
lobula-lobula plate columnar neurons (LLPC) by Strausfeld and Gilbert
(1992), and the later name is now preferred in the literature.

The previous definition for that term was too specific and was seemingly
based on the assumption that the only figure depicting a Lccn2 neuron in
the Fischbach paper was representative of *all* Lccn2 neurons--a claim
that the authors themselves did not do. Likewise, there's nothing to
suggest that the findings of Raghu et al. (2013) regarding the
neurotransmitter applies to all Lccn2 neurons and not to a subset of
them. We update the definition and the corresponding relationships
accordingly.