Fix over-specific term 'S cell'.
The term 'S cell' was defined (both logically and textually) as a
sugar-sensing neuron of the labellum. However, in the literature that
term is used more generally to refer to sugar-sensing neurons in any
taste bristle anywhere in the body, not only in labellar taste bristle.

We update the definitions to match the common meaning in the literature
and remove all informations that are specific to labellar S cells. We
also change the name for a more explicit name.

We add a new term to refer specifically to sugar-sensing neurons of
labellar taste bristles.