Add 'water-sensing neuron'.
The term 'W cell' was defined as referring to water-sensing neurons of
the labellum, but is actually used in the literature to refer to all
water-sensing neurons in taste sensilla anywhere in the body.

Because 'W cell' has been exclusively used in curation with the
labellar-specific meaning, we preserve that meaning but rename the term
to make the labellar-specific bit explicit. We then add a new more
generic term ('water-sensing neuron') to refer to any water-sensing
neuron; that term carries the exact synonym 'W cell'.