Fix over-specific 'L1 GRN'.
The term 'L1 GRN' was defined as referring to low salt-sensing neurons
of the labellum, but is used in the literature to refer to low
salt-sensing neurons in any gustatory sensillum anywhere in the body.

Because 'L1 GRN' has only been used in curation to refer to labellar
neurons, we preserve that meaning, but change the label to make it more
explicit ('low salt-sensing neuron of the labellum').

We then add a more generic term 'low salt-sensing neuron of gustatory
sensillum', carrying the exact synonym 'L1 GRN'. For ease of
classification, we also add a even more generic term ('salt-sensing
neuron of gustatory sensillum') that will encompass both low-salt and
high-salt neurons.