Fix over-specific term 'L2 GRN'.
The term 'L2 GRN' was defined as referring specifically to high
salt-sensing neurons of the labellum, but is used in the literature to
refer to any high salt-sensing neurons in any taste bristle anywhere in
the body.

In curation, the term has been used both with its labellar-specific
meaning and with its more general meaning. We therefore update its label
and definition to give it the more general meaning, and add a new more
specific term to refer to labellar L2 GRNs.