Add 'water-sensing neuron of gustatory sensillum'.
Add an intermediate term between 'water-sensing neuron' and
'water-sensing neuron of the labellum'. That intermediate term is the
one that should carry the 'W cell' synonym. There may be other
water-sensing neurons that do not belong to a taste bristle (e.g. in
larvae), though we don't have any currently in FBbt.