Add 'water-sensing neuron'.
Add a new term to refer to any water-sensing neuron. Rename 'W cell' to
'water-sensing neuron of the labellum' and update its definition to make
it labellar-specific (consistent with usage in curation records). Also
logically define that term as a water-sensing neuron innervating a
labellar bristle. Add expression of water detector ppk28.