Broaden definition of 'L2 GRN'.
Rename 'L2 GRN' into 'high salt/bitter-sensing neuron' and update its
definition to make it refer to any neuron responsive either to high
concentration of salts or to bitter compounds, anywhere in the organism.

Reclassify 'bitter-sensing neuron' under that term.

Update the definition of 'bitter-sensing neuron of the labellum' to
mention that those neurons also respond to high salt.

Add a new term 'high salt-sensing neuron of the labellum' to refer to
the subset of the "L2 GRN" that are only responsive to high salt and not
to bitter compounds (and which are exclusively found in L-type
sensilla).