Rename muscle terms, part 1.
Update muscle terms to make it clear that they refer to the muscle cells
and not the muscles proper: add 'cell' to the labels and update the
definitions accordingly.

This commit concerns most somatic muscle terms apart from the terms for
embryonic/larval muscles and the terms for individual, numbered muscles.