Add term for epithelial cells of the crop.
Add a term representing the section of the gut epithelium lining the
crop ('crop epithelium'), and a term representing the cells that make up
that epithelium ('epithelial cell of crop').

Also update the definition of 'adult foregut epithelium' to mention that
it lines the crop in addition to the esophagus and parts of the
proventriculus.