Add grouping terms for CNS vs PNS neurons.
Add 'CNS neuron' to refer to any neuron of the central nervous system
and conversely 'PNS' neuron for any neuron of the peripheral nervous
system. This is on par with what we already have for glial cells ('CNS
glial cell' and 'PNS glial cell'). Also add 'VNC neuron' for any neuron
of the ventral nerve cord.

Add the 'adult' and 'embryonic/larval' variants for these terms.