Fix generation of mappings.sssom.tsv.
The call to the sort command using the '\t' syntax is not portable and
does not work on the Ubuntu running inside the ODK container (though it
works on native Mac OS).

Replace the '\t' with an actual tab character. This makes the rule a
little bit less clear (it is not obvious than we are using a tab
character as field separator), but at least it works everywhere. And the
source file extension (tsv) should make it clear enough that we are
dealing with tab-separated values.