Remove mapping between FBbt:00005219 and CL:0000409.
FBbt:00005219 (scolopale cell) is mapped to both CL:0000382 (scolopale
cell) and CL:0000409 (scolopidial sheath cell).

CL:0000409 is redundant (scolopidial sheath cell is a synonym for
scolopale cell) and most likely incorrectly classified, as it states
that scolopidial sheath cells are epithelial cells, which they don't
seem to be (see for example descriptions in Klowden 2008).

Remove the mapping to CL:0000409. The CL term will probably need to be
obsoleted and merged with CL:0000382.