Remove bogus mapping to CL:00003828.
FBbt:00005222 (chordotonal neuron) is mapped to a bogus CL ID
(CL:00003828 -- note the 8-digits ID; CL IDs are 7-digits long).

There does not seem to be a term in CL corresponding to chordotonal
neurons. I suspect the intended target ID was CL:0000382 (with an added
terminal '8'), but while that term does refer to a cell that belongs to
a mechanosensory organ (the scolopale cell), it does not refer to the
actual neuron.

Remove the mapping for now.