Update 'cardial cell' and add adult-specific term.
Update the definition of 'cardial cell' to remove the notion of
contractility, since the term is used to refer to all cells that form
the wall of the dorsal vessel, including the cells that form the ostia
and the intracardial valves.

Add an adult-specific 'adult cardial cell' term, on par with the
existing 'embryonic/larval cardial cell'.

Add cell numbers as provided by Lehmacher et al. (2012) [FBrf0219200].