Update various heart-related terms.
Update the definition for the larval heart to remove the mention of 3
chambers (not found in recent descriptions).

Update the definition of 'adult ostium' to mention that there are 5
pairs of ostium in the adult heart, not 4 (the first chamber contains 2
pairs).

Update 'ostiolar valve' to indicate that the cells making up these
structures are the newly added 'ostial cells'.