Remove comment about cuboidal/columnar epithelium mapping.
The mapping between FBbt's 'cuboidal/columnar epithelium' and Uberon's
'simple columnar epithelium' is fine, despite the fact the definition of
the FBbt term does not include the distinction between unilaminar and
stratified epithelia. Given that there are seemingly no stratified
cuboidal/columnar epithelia in Drosophila, it is correct to state this
mapping as "exact".