Fix mappings for muscle cells.
As we currently do not have terms for actual muscles, our muscle cell
terms (both visceral and somatic) should point to the equivalent terms
in CL rather to the muscle terms in Uberon.

Mapping to Uberon muscle terms can be reinstated when we'll have proper
terms for muscles in FBbt.