Fix mapping of 'ventral neurectoderm'.
FBbt's 'ventral neurectoderm' was mapped to Uberon's 'neurectoderm',
despite the FBbt term being more specific. The correct FBbt equivalent
to Uberon's 'neurectoderm' is 'neurogenic region' (which is the parent
class to 'ventral neurectoderm').

FBbt's 'neurogenic region' was already mapped to Uberon's 'insect
neurogenic region', but that term is of dubious use and should probably
be removed from Uberon.