Remove comment about 'exocuticle' mapping.
The mapping between FBbt's 'exocuticle' and the identical term in Uberon
was marked as potentially dubious because the FBbt term is
pupal-specific, which the Uberon term is not. However, given that the
FBbt term is more specific than the Uberon term and not the other way
round, the mapping is still correct (a FBbt exocuticle is a Uberon
exocuticle that is part of some Drosophila, and that also happens to
only exist in the pupae).