Remove mapping of testis sheath.
FBbt's 'testis sheath' refers to a structure containing both an
epithelium and a muscle layer, and therefore should not be mapped to
Uberon's 'epithelium of testis'.

Likewise for 'ovarian sheath'.

There is already a mapping between the sex-neutral terms 'gonadal sheath
epithelium' and 'epithelium of gonad', so that will do for now. If we
add sex-specific terms for 'testis epithelium' and 'ovarian epithelium',
we could map them back to the corresponding Uberon terms.