Test Version ODK 1.3.0
This version is not out yet @Clare72, but if you do not mind, would you be able to take it for a test drive? It has so many changes, that I really want to be sure I didnt break _any_ of your workflows (wether it is prepare_release, or anything else).