Remove mappings to Uberon `ectoderm-derived structure` and similar.
The Uberon term `ectoderm-derived structure` (UBERON:0004121) has a
different scope from FBbt `ectodermal derivative`: it refers to any
structure that ultimately derives from the ectoderm regardless of the
developmental stage that structure exists in; by contrast the FBbt term
strictly refers to *embryonic* derivatives of the ectoderm.

Mapping those two terms is thus wrong, and has dramatic effects when
Uberon and FBbt are merged, as it cause many Drosophila adult structures
to be classified as `part_of some embryo`.

closes #1425