Replace dcterms:contributor by dcelements:contributor when needed.
For contributors who do not have (yet) a formal identifier, replace the
`dcterms:contributor` annotation (which must have a IRI value) by a
`dcelements:contributor` annotation (which does not have such a
requirement).

Those annotations may be switched back later to `dcterms:contributor`
if/when we can get a proper IRI for those contributors.