Add subtypes of `adult fat cell`.
Add `adult abdominal fat cell` and `adult peripheral fat cell` as fat
cells that are part of the `adult abdominal fat mass` and the `adult
peripheral fat mass`, respectively. These two new terms are on par with
the existing `adult pericerebral fat cell`, which refers to fat cells
that are part of the `adult pericerebral fat mass`.

Related to #1501.