Add terms for female cyst cells at specific cyst stages.
We have precise terms to refer to the several stages of female germline
cyst development (2-, 4-, 8-, and 16-cell cyst), but we currently only
have `cyst cell of ovary` to refer to the cells making up those cysts,
regardless of the stage. For the anatomograms, we need terms
representing the specific cyst cells that belong to a particular cyst
stage.