Re-classify `presumptive nurse cell` and `presumptive oocyte`.
The terms `presumptive nurse cell` and `presumptive oocyyte` can be
re-classified under the new `cyst cell of female germline 16-cell cyst`.

Also make `female germline 16-cell cyst` part of the `germarium`,
instead of the larger `ovariole`. As defined, the 16-cell cyst exists
entirely within the germarium -- when it leaves the germarium, it has
already matured as an `egg chamber`.