Relax relation between `hemocyte of leg disc` and `leg disc`.
Do not assert that leg disc hemocytes are fully part of the leg disc, as
it is currently unclear where they are exactly (e.g., at the surface of
the disc or embedded within the disc epithelium?) and whether they can
easily join the circulating pool of hemocytes.

What is clear is that such hemocytes will be present in a dissected leg
disc, which justifies at least the `attached to` relation.