Add a term for the undistended portion of the vas deferens.
The `vas deferens` is divided into a distal, enlarged portion (the
seminal vesicle) and a proximal, undistended portion that connects the
seminal vesicle to the ejaculatory duct. We have a term for the seminal
vesicle, but we are missing a term for the remaining of the vas
deferens, so we add it here. We call it “twisted vas deferens” based on
the description by Miller (1941), though that term does not seem to be
much used (most people seem to refer to that portion by the term “vas
deferens”, excluding the seminal vesicle).