Update the definition of `ejaculatory duct`.
Both the `anterior ejaculatory duct` and the `posterior ejaculatory
duct` are `part_of` the `ejaculatory duct`, but the text definition of
`ejaculatory duct` is actually describing the anterior duct only (from
the vasa deferentia to the ejaculatory bulb).

So we update the definition so that it describes the entire duct.

We also make the `ejaculatory bulb` a `part_of` the entire `ejaculatory
duct`, which is more consistent than having it as a separate structure
(otherwise the `ejaculatory duct` would be made of two disconnected
segments).