Add missing subtypes of spermatocyte cysts and cyst cells.
We have a precise hierarchy representing all the different stages of
male sperm cells (from the primary gonial cell to the mature
spermatozoon), but the corresponding hierarchy for the associated cyst
cells and the cyst themselves is missing several terms.

We add all the missing terms so as to obtain a complete lineage of male
germline cysts:

- spermatogonial cyst
- spermatocyte cyst
  - primary spermatocyte cyst
    - apolar primary spermatocyte cyst
    - polar primary spermatocyte cyst
    - mature primary spermatocyte cyst
  - secondary spermatocyte cyst
- spermatid cyst

with, for each cyst, the corresponding cyst cells and germ cells.