Add germline and cyst cell types of the larval testis.
Add terms for each of the developing stages of germ cells that are found
in the larval testis (from the spermatogonium to the mature primary
spermatocyte), and the accompanying cyst cells.

Contrary to what we have done for the adult testis, here we do not
create terms for each stages of cysts: we use one simple `larval male
germline cyst` instead. Precise subtypes were needed in the adult for
the testis anatomograms, but there are currently no plans to create
anatomograms for larval organs. We’ll defer creating more precise
subtypes until/unless they are needed.