Re-define `somatic cell of larval testis` logically.
Make `somatic cell of larval testis` as equivalent to `somatic cell` and
`part_of some larval testis`, as the term is intended to encompass all
cells of a somatic lineage in the larval testis.

Make `terminal epithelial precursor cell` and `testis pigment cell
precursor` part of `larval testis`, so that they are automatically
classified as `somatic cell of larval testis`.