Unmap `muscle system` from its Uberon counterpart.
In FBbt, `muscle system` is clearly defined as encompassing both somatic
(aka skeletal) and visceral muscles. Its Uberon counterpart is much less
clearly defined, though, but there is a consensus among Uberon editors
that it should refer solely to skeletal muscles, which means our current
mapping is wrong.

In addition, the Uberon term is classified as a `organ system
subdivision` (because it is considered to be the “muscular“ component of
the `musculoskelal system`) while the FBbt term is a `anatomical
system`. The two classifications are incompatible and this results in a
lot of unsatisfiable classes when merging Uberon and FBbt.

Uberon needs to revise its `musculoskeletal system` tree. Until that is
done, for now we unmap our `muscle system` from Uberon’s `musculature of
body`.