Break circular dependency to build flybase_import.owl.
To build the flybase_import.owl component, we need a list of all the
terms ever referred to in the ontology (from which we extract FBgn
"terms"). Until now we re-used $(IMPORTSEED) for that, but this is no
longer possible now that flybase_import.owl is a true component, because
the $(IMPORTSEED) depends on all the components being up-to-date. This
creates a circular dependency where we need flybase_import.owl in order
to generate $(IMPORTSEED), but we need $(IMPORTSEED) in order to
generate flybase_import.owl.

We break that cycle by creating a list of terms (tmp/fbgn_seed.txt)
independently of $(IMPORTSEED), in a rule that depends only on the -edit
file.

This means that imports won't be refreshed prior to generating the
flybase_import.owl component, but that shouldn't be an issue as we don't
expect to find references to FlyBase genes in our GO, RO, or PATO
imports.