Remove the custom preprocessing step.
We have a custom preprocessing step that is used to inject automatically
generated definitions in the ontology.

Here, we update that step to exploit the new preprocessing hook
available in the latest ODK. This removes the need for the `pre_release`
target: a release can now be performed in a single step using the
standard `prepare_release` target.