Define 'aminergic neuron' following literature usage.
In the literature, 'aminergic' is used to refer to a variable set of
neurotransmitters, with sometimes some disagreements (for example, some
authors include GABAergic neurons among the aminergic neurons, some do
not). The most common usage seems to use that term to refer to biogenic
amines only (e.g. excluding GABA), so that is how we define it here.

Trying to follow literage usage implies that we can no longer define the
term logically by referring to GO’s 'primary amine secretion', because
GO is following CHEBI and in CHEBI 'primary amine' refer to a much
larger group of molecules.