Re-classify 'genitalia' as 'anatomical structure'.
The classification of 'genitalia' as an 'anatomical cluster' leads to
unsatisfiability of 'male genital muscle', because that muscle is
'attached to part of' the male genitalia and the 'attached to part of'
relation is range-restricted to a (connected) anatomical structure.

The term "cluster" is probably a bit of a stretch for such a structure
anyway, so we re-classify it as a proper 'anatomical structure'.

Likewise for the similar terms 'analia' and 'terminalia'.