Fix neuroepidermoblast.
Add a proper definition to 'neuroepidermoblast'. Also re-classify as a
'somatic precursor cell' rather than as 'neuroblast', since a
neuroepidermoblast, as the term is understood, is a cell that _may_
become a neuroblast but has not yet done so (and it may become an
epidermoblast instead, thereby never entering the neural lineage).
Assert it as part of a proneural cluster, with a 'proneural cluster
cell' synonym.

Remove the mapping with the equivalent term in CL, as I plan to obsolete
the CL term (CL does not need such a precise term, especially since the
term is actually never used in the literature).

closes #1621