Add distinct classes for DAL and DAL2 neurons.
Add 'dorsal anterior lateral neuron of the protocerebrum' (DAL) and
'dorsal anterior lateral 2 neuron of the protocerebrum' (DAL2). The
former gets most of its definition and relationships from the old
'dorsal anterior lateral neuron of the protocerebrum' (now broadened so
that it refers to any of the two). The latter gets its definition and
relationships from FBrf0252199.