Add another superclass term.
FlyWire's "optic" superclass is slightly broader than the existing
'optic lobe intrinsic neuron', as it is intended to cover the ocellar
ganglia in addition to the optic lobes. So we add a new term to
represent that superclass.

Note the new term is currently pretty badly named: strictly speaking,
they are not related solely to visual "neuropils" (since the ocellar
ganglia are not neuropils), and they are not "intrinsic" (since the
ocellar ganglia are not part of the CNS)...