Fix inconsistencies with 'optic lobe intrinsic neuron'.
Change the logical definition of 'optic lobe intrinsic neuron' so that
it does not imply those neurons are entirely contained within an optic
lobe. Use 'has synaptic IO in region' instead, which fits the way we
define 'intrinsic'.

Remove the 'has soma location' some 'adult optic lobe', so that the term
covers the case of neurons like 'tangential neuron CT1', which arborizes
solely in the optic lobe (and should therefore be considered 'intrinsic'
to the OL) but has its soma in the central brain.

Likewise for 'optic lobe amacrine neuron', which is a direct parent to
'tangential neuron CT1' and therefore should not be said to be 'part of'
the optic lobe.