Replace the old EQWriter tool by a ROBOT plugin.
In the preprocessing pipeline, replace "dot-definitions" using a
specialised ROBOT plugin rather than the unmaintained EQWriter tool.
This greatly simplifies the preprocessing step as everything is done in
a single command (instead of one command to extract the terms with a dot
definition, one command to generate the definitions, one command to
merge the newly generated definitions, and one command to remove the old
dot definitions).