Fix 'adult stomodeal nerve'.
The class 'adult stomodeal nerve' is classified both as part of the PNS
(because it is a nerve) and as part of the CNS (because it is part of
the hypocerebral ganglion).

But the classification as part of the HCG is wrong on two counts:

1) the nerve originates (or terminates, dpeending on your point of view)
   in the vicinity of the HCG, but is not entirely contained within it;

2) in fact the nerve originates not from the HCG but from the corpus
   cardiacum, next to the HCG.

So here we replace the classification as part of the HCG by one stating
that the stomodeal nerve innervates the corpus cardiacum.