Fix 'embryonic/larval frontal nerve'.
The class 'embryonic/larval frontal nerve' is classified both as part of
the PNS (because it is a nerve) and as part of the CNS (because it is
asserted as part of the stomatogastric nervous system, itself a part of
the CNS).

We fix that contradiction by simply removing the classification as part
of the stomatogastric nervous system (SNS). The nerve is already said to
be innervating the frontal ganglion (which is part of the SNS), so it is
still linked to the SNS, albeit indirectly.

Likewise for 'embryonic/larval recurrent nerve', 'larval esophageal
nerve 1', and 'larval esophageal nerve 2'.