Ensure failed workflows are properly reported.
Add a `set -e` at the beginning of the `run.sh` script to ensure that
the script fails if the workflow to run fails. Otherwise we can have a
test workflow that fails but the test suite will still appear to pass.

This is the same fix that is being applied upstream in the ODK.