Unmap Uberon's 'insect dorsal ectoderm derivative'.
In Uberon, an ectoderm "derivative" is necessarily an anatomical
structure, which is a _connected_ anatomical entity. In FBbt, a
derivative can be any kind of anatomical entity, regardless of whether
it is connected or disconnected.

Because of this difference, 'dorsal ectoderm derivative' in FBbt should
not be mapped to 'insect dorsal ectoderm derivative' in Uberon.

None of the other 'derivative' terms in FBbt (e.g. mesodermal
derivative, anterior endoderm derivative, etc.) are mapped to Uberon
terms, and rightly so.