Update the ODK-provided QC workflow.
Update the main test command used by the GitHub Action QC workflow; the
previous command was no longer correct after some of the changes brought
to the preprocessing pipeline.

We currently do _not_ that workflow, but this commit makes sure it is
ready to be use, should we one day decide to switch our QC from Travis
to GitHub Actions.