Add an integration test with Uberon/CL.
Add a GitHub Action workflow that attempts to merge the current FBbt,
Uberon, and CL and check for unsatisfiable classes.

The workflow is set for manual triggering only; we don't want to use it
on every PR because it could fail because of changes that were
introduced on the Uberon/CL side.