Unmap 'gonad' from its Uberon counterpart.
With 'larval ovary' and 'larval testis' being classified as 'gonad',
that term cannot remain mapped to its Uberon counterpart: the mapping
ultimately results in 'larval ovary' and 'ovary' being considered
equivalent, because in Uberon 'ovary' itself is equivalent to 'gonad and
part_of some female reproductive system'. This violates the disjonction
between larval and ovary structures that we have in FBbt, and thus
causes ~200 unsatisfiable classes upon merging Uberon and FBbt (all
classes related to the larval ovary, including the adult ovary which is
said to develops_from the larval ovary).

Likewise for 'larval testis' and 'testis'.

The easiest solution is to unmap FBbt's 'gonad' from Uberon's 'gonad'.
The two terms do not have compatible definitions anyway: the FBbt term
is "the organ of the reproductive system in which the germ cells
reside", whereas the Uberon term is "the reproductive organ that
produces and releases eggs or sperm" -- the Uberon term clearly refers
to mature gonads only, whereas the FBbt term does not have this
restriction.

The sex-specific terms 'ovary' and 'testis' are still properly mapped,
so the loss of information is minimal.