Remove 'alt_id' on obsolete terms.
The 'alt_id' tag carried by a term is interpreted as a way to signify
both

* that the term pointed by the tag is obsolete, and
* that it is replaced by the term carrying the tag.

We had 4 instances where the tag was used to point to terms that were
already marked as obsolete through the use of the `is_obsolete` tag, and
for unclear reasons the simultaneous use of `alt_id` and `is_obsolete`
seemed to disturb the OBO serializer in Protégé, leading it to serialize
the obsoleted terms as OWL in the top-level `owl-axioms:` stanza.

To avoid that, we remove the offending `alt_id` tags. They are not
needed anyway since, as stated, the referenced terms are already marked
as obsolete (and they also have `consider` annotations to point curators
to possible replacement terms).