Remove the intermediate mappings.tsv file.
Do not edit the mappings in a custom, non-SSSOM TSV file. Maintain them
directly in the final SSSOM/TSV format instead. The benefit of using a
custom intermediate format is not enough to justify the additional
complexity.

We can then get rid of the AWK script that was turned the custom TSV
file into a proper SSSOM/TSV file.