Fix classification of somatic muscles of the head.
All somatic muscles of the head are currently classified as hypodermal.
This is incorrect as not all head muscles are connected to the
hypodermis.

We re-classify the following muscles:

* 'larval cardiac outflow muscle': it connects the esophagus to the
  aorta and is therefore not connected to the hypodermis -> we
  re-classify it as a 'larval somatic muscle' and explicitly make it
  'part of' some 'larval head'.

* 'pharyngeal muscle': not all pharyngeal muscles are hypodermal, so
  likewise, we re-classify that term as a 'larval somatic muscle' and
  explicitly make it 'part of' some 'larval head'.

That last change means that children of 'pharyngeal muscle' are no
longer considered hypodermal, so for those children that are actually
hypodermal, we need to explicitly re-assert them as 'embryonic/larval
hypodermal muscle':

* 'ventral pharyngeal muscle',
* 'ventral mouthpart muscle 1',
* 'dorsal protractor muscle'.