Add embryonic "cell of ..." terms.
Embryonic scRNAseq datasets contain several clusters of cells where the
precise cell type is not clearly identified, beyond the fact that they
belong to a particular primordium. Since clusters must be annotated to
cell terms (rather than terms representing the primordia themselves), we
need a bunch a terms that represent the cells of those primordia.