Add 'caudal mesoderm'.
Add 'caudal mesoderm', a distinct section of the mesoderm at the
posterior-most end of the embryo.

The pre-existing term 'longitudinal trunk visceral muscle primordium'
(aka the 'caudal visceral mesoderm' or CVM), is itself a part of the
caudal mesoderm (the part that specifically give rise to the
longitudinal visceral muscles), so we add a link to that effect between
the two terms.