Change the synonym type of "SGP" for 'gonadal sheath proper primordium'.
"Somatic gonadal precursor" (SGP) is used to represent the cells of the
gonadal sheath primordium, not the primordium itself. Now that we have
distinct terms to represent those cells, we can leave the "SGP" synonym
on the term that represents the primordium, but we make it a narrow
synonym rather than an exact synonym.