Make 'adult aortic funnel' part of the 'adult heart'.
In FBbt, the term 'adult heart' refers strictly to the posterior section
of the dorsal vessel, excluding the aorta (which is the anterior section
of said dorsal vessel). Therefore, the 'aortic funnel', which is
textually defined as being part of the aorta, cannot be part of the
'heart'.

So here we make 'adult aortic funnel' part of the 'adult aorta' instead
of the 'adult heart', so as to align the relationship with the textual
definition.

closes #2069