Add missing part_of between FBbt:00059202 and FBbt:00051198.
The `embryonic ventral sensory organ precursor cluster group`
(FBbt:00059198) is made of both the the `embryonic ventral sensory organ
precursor cluster` (FBbt:0005199) and the `embryonic ventral' sensory
organ precursor cluster' (FBbt:00059202), but only the ventral cluster
is linked to the cluster group through a part_of relation; the ventral'
cluster is not.

This commit adds the missing part_of relation.