Add synonyms for Poxn-neurons.
Poxn-neurons show up in several scRNAseq datasets on fly brains.

Unfortunately, they are not a single, well-identified and characterized
cell type, but instead correspond to unclear subsets of several
completely different neurons, found in two different locations: a
protocerebral dorsal cluster (DC) and a deutocerebral ventral cluster
(VC).

According to the best description of Poxn-neurons so far (Minocha et
al., 2017; FBrf0235371):

* the DC Poxn-neurons represent most of the adult ellipsoid body
  R-neuron; however not all R-neurons are DC Poxn-neurons and not all DC
  Poxn-neurons are R-neurons;
* the VC Poxn-neurons represent most of the antennal lobe projection
  neurons (vPN); but again, not all vPNs are VC Poxn-neurons and not all
  VC Poxn-neurons are vPNs;
* VC Poxn-neurons that are not vPNS are ALv1 local interneurons, but not
  all ALv1 local interneurons are VC Poxn-neurons.

Single-cell transcriptomics folks can believe that Poxn expression is
enough to define a cluster as much as they want, there’s no way we can
define a _cell type_ out of that.

So here we simply add related synonyms to the three aforementioned cell
types that overlap the most with “Poxn-neurons”.